Tests/LibWeb/Text/expected/selection-in-focused-editing-host.txt

Range: [object Text] 1 [object Text] 2
document.activeElement: [object HTMLDivElement]
Selection must be the same as above:
Range: [object Text] 1 [object Text] 2
document.activeElement: [object HTMLDivElement]
